Desktop: Refactor settings screen components and improve accessibility labelling
