+++ packages/app-desktop/gui/ConfigScreen/controls/SettingDescription.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+interface Props {
+	text: string;
+	id?: string;
+}
+
+const SettingDescription: React.FC<Props> = props => {
+	return props.text ? <div className='setting-description' id={props.id}>{props.text}</div> : null;
+};
+
+export default SettingDescription;
