+++ packages/app-desktop/gui/ConfigScreen/controls/SettingHeader.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+interface Props {
+	text: string;
+}
+
+const SettingHeader: React.FC<Props> = props => {
+	return (
+		<div className='setting-header'>
+			<label>{props.text}</label>
+		</div>
+	);
+};
+
+export default SettingHeader;
