+++ packages/app-desktop/gui/ConfigScreen/controls/SettingLabel.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+
+interface Props {
+	htmlFor: string|null;
+	text: string;
+}
+
+const SettingLabel: React.FC<Props> = props => {
+	return (
+		<div className='setting-label'>
+			<label htmlFor={props.htmlFor}>{props.text}</label>
+		</div>
+	);
+};
+
+export default SettingLabel;
