+++ packages/app-desktop/gui/ConfigScreen/style.scss
@@ -1,3 +1,5 @@
+@use "./styles/index.scss";
+
 .config-screen-content-wrapper {
 	padding: 24px;
 	overflow: auto;
