+++ packages/app-desktop/gui/ConfigScreen/styles/index.scss
@@ -1,0 +1,4 @@
+
+@use "./setting-description.scss";
+@use "./setting-label.scss";
+@use "./setting-header.scss";
