+++ packages/app-desktop/gui/ConfigScreen/styles/setting-description.scss
@@ -1,0 +1,8 @@
+
+.setting-description {
+	color: var(--joplin-color-faded);
+	font-size: var(--joplin-font-size);
+	font-style: italic;
+	max-width: 70em;
+	margin-top: 5px;
+}
